+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add more events for pre and post write & read operations
 - Add more recent version of LwRB
 - Fix compilation error if CRC mode is disabled
+- Add support for dynamic configuration, to support multiple LwPKT instances in one project
 
 ## v1.2.0